requirements.txt
fastapi
uvicorn[standard]
pymongo==3.11
pymongo[srv]
python-multipart
pydantic-settings